backend/requirements/base.txt
Django>=5.1,<5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.1
django-celery-beat>=2.6
celery[redis]>=5.4
redis>=5.0
psycopg[binary]>=3.1
drf-spectacular>=0.27
Pillow>=10.2
openpyxl>=3.1
gunicorn>=22.0
python-decouple>=3.8
dj-database-url>=2.1
